Move `GoogleListSerializer` to separate file.
